Bug fix, screen will be unclickble after toast(s)
Bug fix screen will be unclickble after toast because the toast div will be still n screen after the toast has ended. This prevented you couldn't click on anything the div was covering. So after the second animation had ended of the div (first popup, second popdown), I added the removal of the div so the divs wouldn't stack on each other.
I came across this bug at a screen resolution of 1024 x 600.